moves trigger to fetchy-boi/trigger & moves trigger function to githubDataStore
